requirements.txt
# Core dependencies
Django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-environ==0.11.2
django-oauth-toolkit==3.0.1
djangorestframework-simplejwt==5.3.1
django-rest-framework-social-oauth2==1.2.0
django-prometheus==2.3.1
django-ratelimit==4.1.0

# Database
psycopg2-binary==2.9.9

# Authentication and OAuth
PyJWT==2.9.0
python3-openid==3.2.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
oauthlib==3.2.2
requests-oauthlib==2.0.0

# API Documentation
drf-yasg==1.21.7
coreapi==2.3.3

# GraphQL
graphene-django==3.2.2
django-graphql-jwt==0.4.0
graphene-file-upload==1.3.0

# Search
elasticsearch==8.15.1
elasticsearch-dsl==8.15.4
django-elasticsearch-dsl==8.0

# Utils
Pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2024.2
whitenoise==6.7.0
gunicorn==23.0.0
cryptography==43.0.1
setuptools>=75.8.0

# Development
black==24.10.0
pre-commit==4.0.1
isort==5.13.2
